Dashboard tabular data and pricing added
The data is now displayed in a user friendly manner, energy usage for that duration is displayed.
This could be updated to show a graph of usage over time but has not been implemented at this stage.
The same pricing component used to display the current balance will be used on the payment page to show the most up to date balance for the individual user.